move ProfileImgForm out to components folder
